# Optimize `array!` and `set!`